Add initial experiments with scenery rendering
It sort of works, but a lot of stuff seems somewhat arbitrary.
Why does it stop drawing when Z gets below 70, for example?

Also need to figure out how VRC does its "squeezing" thing. Perhaps
it's as simple as using those values in the scale for parse-coord?
We would, of course, prefer to keep that state-independent....