Refactor profile updating to not commit every time
Might cause unnecessary disk writes, and just seems obnoxious.
Instead, unless requested otherwise, we just update the profile
in-memory and mark it as dirty, so we can warn when closing
of uncommited changes.

It seems reasonable to commit connections immediately, but we
should probably prompt if the profile is already dirty.