Experiment to handle scrolling "perfectly"
Refs #28

We realize the lazy sequence every time by getting the count
for accurate, wrapped-line-by-line scrolling.

This seems really slow---about 40ms on average---but apparently
that's what the old mechanism took as well. It probably will get
slower linearly with the size of the scrollback, though, which
the other would not. We can probably cache the constructed output,
however, and invalidate it on window resize and new output.

Or, perhaps have two separate caches---one for measurement and one
for the constructed output. When measurement is invalidated, so is
the constructed output, but while measurement is valid, we can
format incoming lines directly into the cache.