Initial bindings and rendering for timers
Refs #38

We use a Swing timer to ensure the UI is redrawn enough
for the timer to be useful. A second may be dropped here and
there, but I think that's acceptable (for now, at least).

Still TODO is to beep when the timer is up, possibly until
the timer is cleared.