Attempt to use lazy parsing for SID/STAR
Refs #43

For some reason, it's getting into a deadlock or a spinloop
or something when trying to inflate with the real data (needed
for expanding navaids to lat/lon, and for handling colors). If
I just pass an empty map for the data, it works fine (with the
above exceptions, of course).

On the bright side, reading both STARs and SIDs with lazy parsing
takes only about 5-7s on this Mac Air, where it took almost a minute
to eagerly read *just* the SIDs on an iMac.