Minor optimization to color parsing in Diagrams
Use (or) instead of the "not-found" value; the latter is
eagerly evaluated, the former is not.

There may be a minor optimization in not using keywords for
color names. Seems like unnecessary overhead; Keyword's hashCode
is not more efficient than the String's, either (in fact, it has
some extra overhead). Still probably minor, but worth trying