Update tracking to reflect real aileron
The existing network code assumed networks could just
toggle the tracking state, but as Vatsim cannot, this
would require the vatsim network to keep track of the
tracked state (or, at least, reference the track ns in
a circular loop) so this is much nicer. Any networks
that *only* support toggling could just use that when
implementing both track! and track-drop!